baekjoon bronze-4 15963

+++ baekjoon/bronze 4/15963/src/Main.java
@@ -2,8 +2,8 @@ import java.util.Scanner;
 public class Main {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
-		long N = sc.nextInt();
-		long M = sc.nextInt();
+		long N = sc.nextLong();
+		long M = sc.nextLong();
 		
 		if(N==M) {
 			System.out.println(1);
